Add error handling and styling
Added try catch block in page.js in order to catch any errors that may
appear when calling the weather api. Added functionality that changes
the color of the page based on how hot the weather is. Added styling to
the page in order to make it look better and keep the information well
organized.